Manga raw club fixed missing pages
manga-raw.club has been changing its website structure so I made the pages images search to not be as specific to the website.
Also made the description search to not be as specific as well so it wont throw null on element.text().
Added R-18 genre to the tag search